+++ src/DefaultCodeTemplate/ApiClientClass/ApiClientInterface.ts
@@ -1,7 +1,6 @@
 import ts from "typescript";
 
-import { Factory } from "../../CodeGenerator";
-import type { CodeGenerator } from "../../types";
+import type { Factory } from "../../factory";
 import type { CodeGeneratorParams } from "../../types/extractSchema";
 
 const httpMethodList: string[] = ["GET", "PUT", "POST", "DELETE", "OPTIONS", "HEAD", "PATCH", "TRACE"];
