+++ src/DefaultCodeTemplate/ApiClientClass/Method.ts
@@ -1,7 +1,6 @@
 import ts from "typescript";
 
-import { Factory } from "../../CodeGenerator";
-import type { CodeGenerator } from "../../types";
+import type { Factory } from "../../factory";
 import type { CodeGeneratorParams } from "../../types/extractSchema";
 import * as MethodBody from "./MethodBody";
