+++ src/factory.ts
@@ -1,0 +1,1 @@
+export * as Factory from "./CodeGenerator/factory";
